requirements.txt
python-dotenv
groq
streamlit
